change audio play to using audio tag
