+++ src/layouts/default.vue
@@ -86,6 +86,7 @@
         </div>
       </div>
     </b-modal>
+    <audio-play ref="audioPlay" />
   </div>
 </template>
 
@@ -93,10 +94,12 @@
 import { db, auth, functions } from "@/plugins/firebase.js";
 import { releaseConfig } from "~/plugins/config.js";
 import DialogBox from "~/components/DialogBox";
+import AudioPlay from "./AudioPlay";
 
 export default {
   components: {
-    DialogBox
+    DialogBox,
+    AudioPlay
   },
   data() {
     return {
@@ -129,9 +132,6 @@ export default {
 
       langPopup: false,
 
-      audioContext: new (window.AudioContext || window.webkitAudioContext)(),
-      pleyedSilent: false,
-      buffer: null
     };
   },
   mounted() {
@@ -182,53 +182,8 @@ export default {
   },
   methods: {
     async enableSound() {
-      if (!this.pleyedSilent) {
-        console.log("default: enableSound");
-        try {
-          const src = this.audioContext.createBufferSource();
-          src.buffer = this.audioContext.createBuffer(1, 1, 22050);
-          src.connect(this.audioContext.destination);
-          src.start(0);
-          console.log("default: silent played");
-
-          this.pleyedSilent = true;
-          this.$store.commit("soundEnable");
-        } catch (e) {
-          console.log(e);
-          Sentry.captureException(e, {
-            tags: {
-              view: "layouts/default",
-              methods: "enableSound"
-            }
-          });
-          console.log("default: layout sound not enabled");
-        }
-      }
+      await this.$refs.audioPlay.enableSound();
     },
-    async play() {
-      if (this.buffer == null) {
-        await this.downloadAudio();
-      }
-      if (this.buffer) {
-        if (this.$store.state.soundOn) {
-          console.log("will play");
-          this.audioContext.decodeAudioData(
-            this.buffer.slice(0),
-            _audioBuffer => {
-              console.log("run internal");
-              const source = this.audioContext.createBufferSource();
-              source.buffer = _audioBuffer;
-              source.connect(this.audioContext.destination);
-              console.log(source);
-              source.start(0);
-            }
-          );
-        } else {
-          console.log("silent order update");
-        }
-      }
-    },
-
     async signout() {
       console.log("signing out", auth.currentUser);
       try {
@@ -272,10 +227,6 @@ export default {
         this.$store.commit("setLang", lang);
       }
     },
-    async downloadAudio() {
-      const res = await fetch(this.$store.state.soundFile);
-      this.buffer = await res.arrayBuffer();
-    },
   },
   beforeCreate() {
     const systemGetConfig = functions.httpsCallable("systemGetConfig");
@@ -305,17 +256,6 @@ export default {
       if (this.$route.query.lang) {
         await this.changeLang(this.$route.query.lang);
       }
-    },
-    async "$store.state.soundFile"() {
-      this.downloadAudio();
-    },
-    async "$store.state.orderEvent"() {
-      await this.play();
-      console.log(`soundEnable = ${this.$store.state.soundEnable}, soundOn=${this.$store.state.soundOn}, soundFile=${this.$store.state.soundFile}`);
-      if (this.buffer == null) {
-        console.log("buffer is null");
-      }
-      console.log(this.$store.state.orderEvent);
     },
     async user() {
       if (this.user) {
